+++ src/core/hle/kernel/client_port.cpp
@@ -14,21 +14,7 @@ namespace Kernel {
 ClientPort::ClientPort() {}
 ClientPort::~ClientPort() {}
 
-Kernel::SharedPtr<ClientPort> ClientPort::CreateForHLE(u32 max_sessions, std::shared_ptr<Service::Interface> hle_interface) {
-    SharedPtr<ClientPort> client_port(new ClientPort);
-    client_port->max_sessions = max_sessions;
-    client_port->active_sessions = 0;
-    client_port->name = hle_interface->GetPortName();
-    client_port->hle_interface = std::move(hle_interface);
-
-    return client_port;
-}
-
 void ClientPort::AddWaitingSession(SharedPtr<ServerSession> server_session) {
-    // A port that has an associated HLE interface doesn't have a server port.
-    if (hle_interface != nullptr)
-        return;
-
     server_port->pending_sessions.push_back(server_session);
     // Wake the threads waiting on the ServerPort
     server_port->WakeupAllWaitingThreads();
